+++ packages/core/src/flows/index.js
@@ -47,6 +47,7 @@ export {
     filterSuggestions,
 } from './recentDestinations.js';
 export { checkRecipientNovelty } from './recipientNovelty.js';
+export { classifySignRisk } from './signRiskClassifier.js';
 export {
     estimateNativeSendFee,
     estimateNativeSendFeeTiers,